+++ frontend/system/components/LoginPage.js
@@ -13,7 +13,7 @@ import {
 } from '@mui/material'
 
 import { useForm } from 'react-hook-form'
-import LogoImg from '../../dist/unibook.png'
+import LogoImg from '../../../dist/unibook.png'
 
 const LoginBox = styled(Stack)`
   display: flex;
